+++ CHANGELOG.rst
@@ -7,11 +7,12 @@ Changelog
 Release date: ``2020-xx-xx``
 
 - `NXPY-192 <https://jira.nuxeo.com/browse/NXPY-192>`__: Add support for Python 3.10
+- `NXPY-193 <https://jira.nuxeo.com/browse/NXPY-193>`__: Fix thread-safety in uploads and workflows
 
 Technical changes
 -----------------
 
--
+- Removed utils.py::``SwapAttr``
 
 3.1.1
 -----